Minor cleanup:
- Replace Int with long.
- Rename ShowUiModel to TvShow
- Update imports

+++ shared/domain/episodes/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/api/EpisodesInteractor.kt
@@ -22,7 +22,7 @@ class EpisodesInteractor constructor(
 }
 
 data class EpisodeQuery(
-    val tvShowId: Int,
-    val seasonId: Int,
+    val tvShowId: Long,
+    val seasonId: Long,
     val seasonNumber: Int
 )
