+++ shared/domain/episodes/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/implementation/EpisodeRepositoryImpl.kt
@@ -21,8 +21,8 @@ class EpisodeRepositoryImpl(
 ) : EpisodeRepository {
 
     override fun observeSeasonEpisodes(
-        tvShowId: Int,
-        seasonId: Int,
+        tvShowId: Long,
+        seasonId: Long,
         seasonNumber: Int
     ): Flow<Resource<List<EpisodesBySeasonId>>> = networkBoundResource(
         query = { episodesCache.observeEpisode(seasonId) },
@@ -35,7 +35,7 @@ class EpisodeRepositoryImpl(
 
     private fun mapAndCache(
         response: SeasonResponse,
-        seasonId: Int
+        seasonId: Long
     ) {
         val episodeEntityList = response.toEpisodeCacheList()
